+++ client-app/vite.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import { resolve } from 'path'
 
+/// <reference types="vitest" />
+
 // https://vitejs.dev/config/
 export default defineConfig(async () => ({
   plugins: [vue()],
@@ -30,5 +32,9 @@ export default defineConfig(async () => ({
     minify: !process.env.TAURI_DEBUG ? 'esbuild' : false,
     // produce sourcemaps for debug builds
     sourcemap: !!process.env.TAURI_DEBUG,
+  },
+  test: {
+    environment: 'jsdom',
+    globals: true,
   },
 }))